Splitting most of route logic into a controller
