+++ src/pages/currency-converter.tsx
@@ -1,12 +1,14 @@
-import React, { SyntheticEvent, useEffect, useState } from 'react';
-import * as TE from 'fp-ts/TaskEither'
-import * as E from 'fp-ts/lib/Either';
+import React, { useState } from 'react';
 import * as RD from '@devexperts/remote-data-ts';
 import { pipe } from 'fp-ts/lib/function';
 
-import CurrencyService, { CurrencyMatrix } from 'services/currency';
+import { USDRatioFromOption, RDRatioResponse } from 'controllers/CurrencyController';
+import type { CurrencyMatrix } from 'services/currency';
 import Layout from 'components/Layout'
 import styles from 'styles/Home.module.scss'
+import { string } from 'io-ts';
+
+export { getStaticProps } from 'controllers/CurrencyController';
 
 type Props = {
 	topCurrencies: CurrencyMatrix | null,
@@ -14,47 +16,16 @@ type Props = {
 	error: string | null,
 }
 
-type RatioResponse = { currency: string, ratio: number };
-
-export async function getStaticProps() {
-	const currencyRes = await CurrencyService.fetchCurrenciesListByRelevance();
-
-	console.log('STATIC_PROPS_DATA', currencyRes);
-
-	return {
-		props: {
-			error: E.isLeft(currencyRes) ? currencyRes?.left.message : null,
-			otherCurrencies: E.isRight(currencyRes) ? currencyRes?.right?.left : null,
-			topCurrencies: E.isRight(currencyRes) ? currencyRes?.right?.right : null,
-		},
-	}
-}
-
 export default function CurrencyConverter({ topCurrencies, otherCurrencies, error }: Props) {
-	const [selectedSymbol, setSelectedSymbol] = useState<RD.RemoteData<Error, RatioResponse>>(RD.initial);
+	const [selectedSymbol, setSelectedSymbol] = useState<RDRatioResponse>(RD.initial);
 	console.log('StaticProps: 💵 Currency Converter => ', { topCurrencies, otherCurrencies, error });
 
-	const fetchSelectedOption = (ev: SyntheticEvent<HTMLSelectElement>) => {
-		const currencySymbol = ev.target.value;
-		setSelectedSymbol(RD.pending);
-
-		return pipe(currencySymbol,
-			CurrencyService.fetchCurrencyToUSDRatio,
-			TE.map(ratio => ({ currency: currencySymbol, ratio })),
-			TE.bimap(RD.failure, RD.success),
-			TE.fold(
-				(err) => TE.fromIO(() => setSelectedSymbol(err)),
-				(res) => TE.fromIO(() => setSelectedSymbol(res))
-			),
-		)();
-	}
-
-	const generateOptions = (currencies: Props['topCurrencies'], optgroup?: string) => {
+	const OptionsSection = ({title, currencies}: { title: string, currencies: CurrencyMatrix | null }) => {
 		if (!currencies?.length) return null;
 
 		const OptWrapper = ({ children }: { children: React.ReactNode[] }) =>
-			optgroup
-				? (<optgroup label={optgroup}>{children}</optgroup>)
+			title
+				? (<optgroup label={title}>{children}</optgroup>)
 				: (<>{children}</>)
 
 		return (
@@ -70,10 +41,12 @@ export default function CurrencyConverter({ topCurrencies, otherCurrencies, erro
 
 	const CurrencySelector = () => {
 		return (
-			<select onChange={fetchSelectedOption}>
+			<select
+				onChange={(ev) => USDRatioFromOption(ev?.target?.value, setSelectedSymbol)}
+			>
 				<option disabled selected>-- Select a currency -- </option>
-				{generateOptions(topCurrencies, 'Most exchanged')}
-				{generateOptions(otherCurrencies, 'A-Z')}
+				<OptionsSection title='Most exchanged' currencies={topCurrencies} />
+				<OptionsSection title='A-Z' currencies={otherCurrencies} />
 			</select>
 		)
 	};
